Different message when breaking known wand of nothing
The existing message "But nothing else happens..." implies
that something could have happened or the wand had some
power that doesn't get used when breaking. If the player
is actually breaking an identified wand of nothing, it's
obvious that nothing will happen. In this case, the
message is changed to "Predictably, nothing happens."