Fix: ghoul and ettin zombie mis-ordered in monster list
An assumption of monster generating code is that monsters within a class
appear in increasing order of difficulty. This wasn't the case in the Z
class which went: human zombie (diff 5) -> ettin zombie (diff 7) ->
ghoul (diff 5).

As a result, ghouls would not randomly generate when they should have at
certain level difficulties (where a ghoul is weak enough to generate but
an ettin zombie would be too strong).

Fix this by swapping the two monsters, which makes the difficulties in
the Z class monotonically increase again.